Total re-write of Spektrum support.  This is a DRAFT and I will optimize it further. It has been tested on a Mega2560 only.
Removed the special interrupt handler and re-wrote support to use the new Serial (no dot) routines.  This eliminates a lot of overhead.  No more special interrupt handler, no more calls to micros(), instead it uses the main loops timing, no extra buffers for the serial data, and much more. 


Also fixed a problem with the new SerialOpen.  It was setting the TX interrupt bits for serial 1, 2, and 3, yet there were no transmit interrupt handlers defined.  This caused an immediate halt when any of these ports was opened.  They are now defined as receive only.  If we ever need them to transmit, we will need to define interrupt handlers and add the bits back into the open.